Enable exceptions and general cleanup
Let's handle exceptions in a way that if they are thrown it wouldn't
be the end of the world to just crash, if it was let's say embedded system
or something like that. In normal runtime situations we shouldn't be
throwing exceptions really. Exceptions is the only way to handle errors
in constructors, and makes construction of objects much much cleaner
so I want to use them for that. Let's just defer construction that can
fail to as early in the program as possible

Also noticed that windows release build is crashing with wine...
Debug build seems to work fine